UNI-388: Move anomaly constants and text mapping to common-utils
